rename tests in mmtbx to follow libtbx convention (current naming is picked up by pytest discovery, with these tests being incompatible with pytest, causing subsquent failures). change documentation and import statements involving these test modules to compensate.
