+++ mmtbx/run_tests.py
@@ -161,10 +161,10 @@ general_tests = [
   "$D/geometry_restraints/tst_reference_coordinate2.py",
   "$D/geometry_restraints/tst_c_beta_restraints.py",
   "$D/geometry_restraints/torsion_restraints/tst_torsion_ncs.py",
-  "$D/conformation_dependent_library/test_cdl.py",
-  "$D/conformation_dependent_library/test_cdl_esd.py",
-  "$D/conformation_dependent_library/test_rdl.py",
-  "$D/conformation_dependent_library/test_hpdl.py",
+  "$D/conformation_dependent_library/tst_cdl.py",
+  "$D/conformation_dependent_library/tst_cdl_esd.py",
+  "$D/conformation_dependent_library/tst_rdl.py",
+  "$D/conformation_dependent_library/tst_hpdl.py",
   "$D/conformation_dependent_library/tst_pH_mechanism.py",
   "$D/regression/tst_find_ss_structure.py",
   "$D/regression/tst_fix_rotamer_outliers.py",
@@ -350,7 +350,7 @@ general_tests = [
   "$D/building/tst_make_library.py",
   "$D/regression/real_space_refine_chain/tst_00.py",
   "$D/regression/real_space_refine_chain/tst_01.py",
-  "$D/conformation_dependent_library/test_omega_cdl.py",
+  "$D/conformation_dependent_library/tst_omega_cdl.py",
   "$D/regression/fix_cablam/tst_basic_cl_operations.py",
   "$D/regression/fix_cablam/tst_single_outliers_surroundings.py",
   "$D/regression/fix_cablam/tst_one_resid_rotation.py",
